ansbl: install vagrant via .deb in 'mmackey' role
Implementation Note: I used files in /tmp to store the SHA-256 sums
for the vagrant .deb and jetbrains tarball.  This was so that I could
make sure the tasks that download the SHA-256 sums don't always show
up as 'changed'.  I originally was just storing the checksums within
variables but that didn't allow me to take advantage of the 'creates'
parameter of the 'shell' module.

Also, I had tried just using 'tee' in the 'shell' command to store
the output in a variable AND in a file in /tmp so I could take advantage
of 'creates' and also won't have to do the following:

```
get_url:
  xxxxx
vars:
  sha256_var: {{ lookup('file'... }}
```

However, when the /tmp file already exists, the 'shell' command never
runs. Thus, the '.stdout' value that would get stored in a variable with
'register' to use in get_url is not a valid checksum (it's just some
warning message) so the get_url.checksum validation throws an error on
every run when the /tmp file already exists.