k8s: moved set_node_ip code to own task file
Goal is that this task can be used in a separate play to configure
kubelet on a new VM that has kubernetes pre-installed.  This needs to be
done separately b/c I want to have a pre-installed Kubernetes VM image
to reduce the time needed to spin up a cluster, and the IP address
during the pre-installation will not match that IP address when the
image is being run for real.