packages: role w/ helper tasks to install pkgs
This role is intended initially to simply hold task files that automate
tasks repeated throughout other roles for installing packages from
things like remote debs or archives.  I don't have time to create an
ansible module for this and note sure how to just have simple task files
outside of roles in order to achieve this.

The way I am using this is probably some sort of hack but I needed to
eliminate the repetitive code and didn't have time to research a bette
way.

This 1st task contains a means to install a remote .deb file, which is
commonly done for tools that are not in the Ubuntu repos.